+++ tests/deprecated_api/test_remove_1-4.py
@@ -30,20 +30,6 @@ from tests.helpers import BoringModel
 from tests.helpers.runif import RunIf
 
 
-def test_v1_4_0_deprecated_trainer_attributes():
-    with pytest.deprecated_call(match="will be removed in v1.4."):
-        trainer = Trainer()
-        _ = trainer.accelerator_backend
-    assert trainer.accelerator == trainer.accelerator_backend
-
-
-def test_v1_4_0_deprecated_trainer_methods():
-    with pytest.deprecated_call(match='will be removed in v1.4'):
-        trainer = Trainer()
-        _ = trainer.get_model()
-    assert trainer.get_model() == trainer.lightning_module
-
-
 def test_v1_4_0_deprecated_imports():
     _soft_unimport_module('pytorch_lightning.utilities.argparse_utils')
     with pytest.deprecated_call(match='will be removed in v1.4'):
@@ -60,87 +46,6 @@ def test_v1_4_0_deprecated_imports():
     _soft_unimport_module('pytorch_lightning.utilities.xla_device_utils')
     with pytest.deprecated_call(match='will be removed in v1.4'):
         from pytorch_lightning.utilities.xla_device_utils import XLADeviceUtils  # noqa: F811 F401
-
-
-def test_v1_4_0_deprecated_trainer_device_distrib():
-    """Test that Trainer attributes works fine."""
-    trainer = Trainer()
-    trainer.accelerator_connector._distrib_type = None
-    trainer.accelerator_connector._device_type = None
-
-    with pytest.deprecated_call(match='deprecated in v1.2 and will be removed in v1.4'):
-        trainer.on_cpu = True
-    with pytest.deprecated_call(match='deprecated in v1.2 and will be removed in v1.4'):
-        assert trainer.on_cpu
-
-    with pytest.deprecated_call(match='deprecated in v1.2 and will be removed in v1.4'):
-        trainer.on_gpu = True
-    with pytest.deprecated_call(match='deprecated in v1.2 and will be removed in v1.4'):
-        assert trainer.on_gpu
-
-    with pytest.deprecated_call(match='deprecated in v1.2 and will be removed in v1.4'):
-        trainer.on_tpu = True
-    with pytest.deprecated_call(match='deprecated in v1.2 and will be removed in v1.4'):
-        assert trainer.on_tpu
-    trainer.accelerator_connector._device_type = None
-    with pytest.deprecated_call(match='deprecated in v1.2 and will be removed in v1.4'):
-        trainer.use_tpu = True
-    with pytest.deprecated_call(match='deprecated in v1.2 and will be removed in v1.4'):
-        assert trainer.use_tpu
-
-    with pytest.deprecated_call(match='deprecated in v1.2 and will be removed in v1.4'):
-        trainer.use_dp = True
-    with pytest.deprecated_call(match='deprecated in v1.2 and will be removed in v1.4'):
-        assert trainer.use_dp
-
-    with pytest.deprecated_call(match='deprecated in v1.2 and will be removed in v1.4'):
-        trainer.use_ddp = True
-    with pytest.deprecated_call(match='deprecated in v1.2 and will be removed in v1.4'):
-        assert trainer.use_ddp
-
-    with pytest.deprecated_call(match='deprecated in v1.2 and will be removed in v1.4'):
-        trainer.use_ddp2 = True
-    with pytest.deprecated_call(match='deprecated in v1.2 and will be removed in v1.4'):
-        assert trainer.use_ddp2
-
-    with pytest.deprecated_call(match='deprecated in v1.2 and will be removed in v1.4'):
-        trainer.use_horovod = True
-    with pytest.deprecated_call(match='deprecated in v1.2 and will be removed in v1.4'):
-        assert trainer.use_horovod
-
-
-def test_v1_4_0_deprecated_metrics():
-    from pytorch_lightning.metrics.functional.classification import stat_scores_multiple_classes
-    with pytest.deprecated_call(match='will be removed in v1.4'):
-        stat_scores_multiple_classes(pred=torch.tensor([0, 1]), target=torch.tensor([0, 1]))
-
-    from pytorch_lightning.metrics.functional.classification import iou
-    with pytest.deprecated_call(match='will be removed in v1.4'):
-        iou(torch.randint(0, 2, (10, 3, 3)), torch.randint(0, 2, (10, 3, 3)))
-
-    from pytorch_lightning.metrics.functional.classification import recall
-    with pytest.deprecated_call(match='will be removed in v1.4'):
-        recall(torch.randint(0, 2, (10, 3, 3)), torch.randint(0, 2, (10, 3, 3)))
-
-    from pytorch_lightning.metrics.functional.classification import precision
-    with pytest.deprecated_call(match='will be removed in v1.4'):
-        precision(torch.randint(0, 2, (10, 3, 3)), torch.randint(0, 2, (10, 3, 3)))
-
-    from pytorch_lightning.metrics.functional.classification import precision_recall
-    with pytest.deprecated_call(match='will be removed in v1.4'):
-        precision_recall(torch.randint(0, 2, (10, 3, 3)), torch.randint(0, 2, (10, 3, 3)))
-
-    from pytorch_lightning.metrics.functional.classification import auc
-    with pytest.deprecated_call(match='will be removed in v1.4'):
-        auc(torch.rand(10, ).sort().values, torch.rand(10, ))
-
-    from pytorch_lightning.metrics.functional.classification import auroc
-    with pytest.deprecated_call(match='will be removed in v1.4'):
-        auroc(torch.rand(10, ), torch.randint(0, 2, (10, )))
-
-    from pytorch_lightning.metrics.functional.classification import multiclass_auroc
-    with pytest.deprecated_call(match='will be removed in v1.4'):
-        multiclass_auroc(torch.rand(20, 5).softmax(dim=-1), torch.randint(0, 5, (20, )), num_classes=5)
 
 
 class CustomDDPPlugin(DDPSpawnPlugin):